% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_imagej_results.R
\name{getDeltaLeafArea}
\alias{getDeltaLeafArea}
\alias{makeDeltaLeafAreaTable}
\title{Calculate differences in leaf area between t and t+1 from `leaf_area` table.}
\usage{
makeDeltaLeafAreaTable(leaf_area)
}
\arguments{
\item{leaf_area}{data.table generated by `makeLeafAreaTable`.}
}
\value{
data.table of time differences in leaf area.
}
\description{
Calculate differences in leaf area between t and t+1 from `leaf_area` table.
}
