% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_imagej_results.R
\name{getPeaks}
\alias{getPeaks}
\title{Detect peaks and return position information for single ROI.}
\usage{
getPeaks(leaf_area_diff_sub, thr_start = 0.2, thr_end = 0.05)
}
\arguments{
\item{leaf_area_diff_sub}{A part (single ROI) of data.table generated by `makeDeltaLeafAreaTable`.}

\item{thr_start}{Numeric. Threshold value of area diff to detect start of the peak.}

\item{thr_end}{Numeric. Threshold value of area diff to detect start of the peak.}
}
\value{
`data.table` with 4 column: start,end,area,interval. Returns NULL when no peak is detected.
}
\description{
Detect peaks and return position information for single ROI.
}
