% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_imagej_results.R
\name{getPeaksAllRoi}
\alias{getPeaksAllRoi}
\title{Detect peaks and return position information for all ROIs.}
\usage{
getPeaksAllRoi(leaf_area_diff, thr_start = 0.2, thr_end = 0.05)
}
\arguments{
\item{leaf_area_diff}{data.table generated by `makeDeltaLeafAreaTable`.}

\item{thr_start}{Numeric. Threshold value of area diff to detect start of the peak.}

\item{thr_end}{Numeric. Threshold value of area diff to detect start of the peak.}
}
\description{
Detect peaks and return position information for all ROIs.
}
